+++ R/lengths.R
@@ -2,11 +2,13 @@
 #'
 #' `bm_widths()` returns the widths of the bitmaps while
 #' `bm_heights()` returns the heights of the bitmaps.
+#' `bm_widths()` and `bm_heights()` are S3 generic functions.
 #'
-#' @inheritParams bm_clamp
+#' @param x Bitmap object.
 #' @param unique Apply [base::unique()] to the returned integer vector.
+#' @param ... `bm_heights.default()` and `bm_widths.default()` pass this to `as_bm_pixmap(x, ...)`.
 #' @return A integer vector of the relevant length of each
-#'         of the `bm_bitmap()` objects in `x`.
+#'         of the bitmap objects in `x`.
 #'         If `unique` is `TRUE` then any duplicates will have been removed.
 #' @examples
 #'   font_file <- system.file("fonts/spleen/spleen-8x16.hex.gz", package = "bittermelon")
@@ -15,12 +17,67 @@
 #'   bm_heights(font) # every glyph in the font is 16 pixels high
 #' @rdname lengths
 #' @export
-bm_widths <- function(bm_object, unique = TRUE) {
-    stopifnot(is_bm_object(bm_object))
-    if (is_bm_bitmap(bm_object))
-        nc <- ncol(bm_object)
+bm_heights <- function(x, ...) {
+    UseMethod("bm_heights")
+}
+
+#' @rdname lengths
+#' @export
+bm_heights.bm_matrix <- function(x, ...) {
+    nrow(x)
+}
+
+#' @rdname lengths
+#' @export
+bm_heights.bm_list <- function(x, unique = TRUE, ...) {
+    nr <- vapply(x, nrow, integer(1L))
+    if (unique)
+        base::unique(nr)
     else
-        nc <- sapply(bm_object, ncol)
+        nr
+}
+
+#' @rdname lengths
+#' @export
+bm_heights.default <- function(x, ...) {
+    nrow(as_bm_pixmap(x, ...))
+}
+
+#' @rdname lengths
+#' @export
+`bm_heights.magick-image` <- function(x, ...) {
+    stopifnot(requireNamespace("magick", quietly = TRUE))
+    magick::image_info(x)$height
+}
+
+#' @rdname lengths
+#' @export
+bm_heights.nativeRaster <- function(x, ...) {
+    nrow(x)
+}
+
+#' @rdname lengths
+#' @export
+bm_heights.raster <- function(x, ...) {
+    nrow(x)
+}
+
+#' @rdname lengths
+#' @export
+bm_widths <- function(x, ...) {
+    UseMethod("bm_widths")
+}
+
+#' @rdname lengths
+#' @export
+bm_widths.bm_matrix <- function(x, ...) {
+    ncol(x)
+}
+
+#' @rdname lengths
+#' @export
+bm_widths.bm_list <- function(x, unique = TRUE, ...) {
+    nc <- vapply(x, ncol, integer(1L))
     if (unique)
         base::unique(nc)
     else
@@ -29,14 +86,25 @@ bm_widths <- function(bm_object, unique = TRUE) {
 
 #' @rdname lengths
 #' @export
-bm_heights <- function(bm_object, unique = TRUE) {
-    stopifnot(is_bm_object(bm_object))
-    if (is_bm_bitmap(bm_object))
-        nr <- nrow(bm_object)
-    else
-        nr <- sapply(bm_object, nrow)
-    if (unique)
-        base::unique(nr)
-    else
-        nr
+bm_widths.default <- function(x, ...) {
+    ncol(as_bm_pixmap(x, ...))
+}
+
+#' @rdname lengths
+#' @export
+`bm_widths.magick-image` <- function(x, ...) {
+    stopifnot(requireNamespace("magick", quietly = TRUE))
+    magick::image_info(x)$width
+}
+
+#' @rdname lengths
+#' @export
+bm_widths.nativeRaster <- function(x, ...) {
+    ncol(x)
+}
+
+#' @rdname lengths
+#' @export
+bm_widths.raster <- function(x, ...) {
+    ncol(x)
 }